requirements.txt
ecapture
python-escpos
python-twitter
